address comments by @trustin
