+++ core/src/main/java/com/linecorp/armeria/common/logging/RequestContextExporterBuilder.java
@@ -129,10 +129,10 @@ public final class RequestContextExporterBuilder {
     /**
      * Specifies a prefix of the default export group.
      */
-    public void setDefaultExportPrefix(String defaultExportPrefix) {
-        requireNonNull(defaultExportPrefix, "defaultExportPrefix");
-        checkArgument(!defaultExportPrefix.isEmpty(), "defaultExportPrefix must not be empty");
-        defaultExportGroupBuilder.setPrefix(defaultExportPrefix);
+    public void prefix(String prefix) {
+        requireNonNull(prefix, "prefix");
+        checkArgument(!prefix.isEmpty(), "prefix must not be empty");
+        defaultExportGroupBuilder.prefix(prefix);
     }
 
     /**
